+++ test/cli/run.spec.ts
@@ -334,6 +334,9 @@ Options:
                                    This option conflicts with the \`-t/--test\`
                                    and \`-T/--test-file\` options unless
                                    \`-o/--output\` is also specified
+  -S, --start-rule <rule>          When testing, use the given rule as the
+                                   start rule.  If this rule is not in the
+                                   allowed start rules, it will be added.
   -t, --test <text>                Test the parser with the given text,
                                    outputting the result of running the parser
                                    instead of the parser itself. If the input
@@ -931,6 +934,16 @@ Options:
       expected: "'boo'\n",
     });
 
+    // Start rule
+    await exec({
+      args: ["-t", "2", "-S", "bar"],
+      stdin: `\
+foo='1' { throw new Error('bar') }
+bar = '2'
+`,
+      expected: "'2'\n",
+    });
+
     const grammarFile = path.join(__dirname, "..", "..", "examples", "json.pegjs");
     const testFile = path.join(__dirname, "..", "..", "package.json");
 
@@ -948,9 +961,9 @@ Options:
 
     await exec({
       args: ["-t", "boo", "-T", "foo"],
-      errorCode: "peggy.invalidArgument",
+      errorCode: "commander.conflictingOption",
       exitCode: 1,
-      error: "The -t/--test and -T/--test-file options are mutually exclusive.",
+      error: "error: option '-T, --test-file <filename>' cannot be used with option '-t, --test <text>'",
     });
 
     await exec({